requirements.txt
streamlit
python-dotenv
pydantic
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
langchain-text-splitters
faiss-cpu
sentence-transformers
pypdf
